Refactored Education frontend components
